Move reinforcement functions to separate module
